requirements.txt
requests
openai
python-docx
